+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Use Common Dyes Tags for Colored Sand Recipes
 - Add Stripped Woods and Logs to new Tags (Added in NeoForge 21.1.71 and Fabric API 0.106.0)
 - Override getShadowRadius in Entity Renderers instead of using render method
+- Make Entity Model Classes Package-Private, remove pointless override of getAnimation Method
 
 # 2.1.4 (Forge Only)
 - Require Forge 52.0.20 as a minimum due to the addition of new Forge tags and implementation of Common Tags